Large arbor update.
Arbor is now a base class without spatial data,
and is extended to handle use case specific data such as
node locations and masks
